Extract dev Linux profile policy
